pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memguard"
version = "0.1.0"
description = "Agent Memory Security — State Firewall for AI Agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "SafePatch Team"},
]
keywords = ["ai", "agent", "security", "memory", "safety", "langchain"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pynacl>=1.5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
langchain = [
    "langchain-core>=0.3.0",
]
detection = [
    "sentence-transformers>=3.0.0",
    "numpy>=1.24.0",
]
dashboard = [
    "streamlit>=1.30.0",
    "pandas>=2.0.0",
    "plotly>=5.18.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
all = [
    "memguard[langchain,detection,dashboard,dev]",
]

[project.urls]
Homepage = "https://github.com/safepatch/memguard"
Documentation = "https://docs.safepatch.dev"

[tool.setuptools.packages.find]
include = ["memguard*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
